c v 5.0.1

+++ app/src/main/java/com/nationalplasticfsm/features/viewAllOrder/ViewAllOrderListRecyclerViewAdapter.java
@@ -140,7 +140,9 @@ public class ViewAllOrderListRecyclerViewAdapter extends RecyclerView.Adapter<Vi
                 if (mViewAllOrderListEntityArray.get(position).getAmount().contains("\u20B9"))
                     attendanceFragmentViewHolder.ordered_amount_tv.setText(mViewAllOrderListEntityArray.get(position).getAmount());
                 else {
-                    String finalAmount = String.format("%.2f", Float.parseFloat(mViewAllOrderListEntityArray.get(position).getAmount()));
+                    //String finalAmount = String.format("%.2f", Float.parseFloat(mViewAllOrderListEntityArray.get(position).getAmount()));
+                    //mantis id 26274
+                    String finalAmount = String.format("%.2f", Double.parseDouble(mViewAllOrderListEntityArray.get(position).getAmount()));
                     attendanceFragmentViewHolder.ordered_amount_tv.setText(context.getString(R.string.rupee_symbol) +
                             " " + finalAmount);
                 }
